Add "Full Roster" report
This serves as a proof-of-concept for PDF report generation.

The old bamru.net used a native Ruby PDF-writing library with its own
template language (prawn), and I didn't spot a Python equivalent, so
simply reusing the existing templates seems infeasible. I chose the
alternative approach of generating HTML and rendering it to PDF using
the `pdfkit` python library, which wraps `wkhtmltopdf`. This has the
advantage of using the same templating and markup languages that we all
already know and love(?).

The main drawback I've encountered is that wkhtmltopdf is a new external
program that needs to be installed, and to get reasonable rendering we
need a version newer than the one in the xenial apt repository. The
project publishes a package, but this still adds more steps to our
setup.