Use html input type="date"/"time" in event edit
I spent some time looking at javascript libraries for date/time picking,
but then I discovered that these input types exist, and modern browsers
have totally reasonable behaviour for them. In particular, on mobile,
this is probably ideal behaviour: at least with Chrome on Android, the
these invoke the system's standard date and time pickers!

This commit also includes form logic to validate that the finish isn't
before the start for this particular usage. It's possible to avoid this
check via the REST API, and I can't find a good way to implement this
validation once so that it applies everywhere.