Move to using internal event IDs for gcal events
There is a long-standing problem where our google calendar proliferates
multiple copies of the same event. I think the most likely cause of this is
transient failures when deleting the gcal version of an event before
re-creating it with updated info (though other failures may also be possible).

This addresses that problem by re-using internal event IDs for events on the
google calendar, and relying on google calendar to maintain uniqueness.

Note that this may not be a full solution, since the google calendar API docs
say: "Due to the globally distributed nature of the system, we cannot guarantee
that ID collisions will be detected at event creation time." So maybe we will
still end up with some duplicate events. But, I'm optimistic that this way will
be better than before.